Add files via upload
I'm not sure if I understood the dist d correctly for the bbox type. Does it add in everything that falls within a square of dimension 2d x 2d with center as the given center? If yes, then I made different functions which put together will generate n splits for the whole map. I used the default, uncleaned database for now. I haven't debugged this, yet. Let's talk more about this tomorrow to finalize it.